Moved `plugin` module to `plugin.rs`
